requirements.txt
0x-contract-addresses==3.0.0
0x-contract-artifacts==3.0.0
0x-json-schemas==2.1.0
0x-order-utils==4.0.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
bitarray==2.8.0
cbor2==5.4.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.2
cytoolz==0.12.2
dacite==1.8.1
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.1
django-ckeditor==6.7.0
django-ipware==5.0.0
django-js-asset==2.1.0
django-sendgrid-v5==1.2.2
django-vote==2.4.0
django-widget-tweaks==1.5.0
dlib==19.24.2
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
face-recognition==1.3.0
face-recognition-models==0.3.0
frozenlist==1.4.0
geoip2==4.7.0
hexbytes==0.3.1
html2image==2.0.4.3
idna==3.4
imgkit==1.2.3
imutils==0.5.4
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.2.0
maxminddb==2.4.0
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==0.4.4
mysqlclient==2.2.0
netaddr==0.8.0
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
parsimonious==0.8.1
pilgram==1.2.1
Pillow==10.0.0
protobuf==3.20.3
psycopg2==2.9.6
pycparser==2.21
pycryptodome==3.18.0
pyee==9.1.1
PyJWT==2.8.0
pymerkle==2.0.2
pyrsistent==0.19.3
python-dotenv==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytracking2==0.4.5
pytz==2022.7.1
qrcode-artistic==2.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rlp==2.0.1
segno==1.5.2
sendgrid==6.10.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stringcase==1.2.0
thirdweb-contract-wrappers==2.0.4
thirdweb-eth-account==0.6.6
thirdweb-sdk==3.1.0
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
varint==1.0.2
w3storage==0.0.1
web3==5.27.0
webcam==1.30
websocket-client==1.6.3
websockets==9.1
wrapt==1.15.0
yarl==1.9.2
